re-structure files to reduce file size and complexity
